Removed map layer style hook and colocated it within the map layers file.
